requirements.txt
PyYAML==3.12
urllib3==1.13.1